Print debug messages with `write(2)`
On my system, calling printf in the allocator causes
issues with the heap. To circumvent it, we use a fixed
size buffer, vsnprintf(3) and write(2) for printing.

This way, the allocator can show what it's doing. Before,
I used GDB to inspect its activity. However, there is a
bug *somewhere* (likely in my own code) that makes `ls`
segfault when I run it with `LD_PRELOAD=./malloc.so` in
GDB. Running `LD_PRELOAD=./malloc.so ls` outside of GDB works.

Now that we can print from inside the allocator, GDB is
out of the loop for testing the allocator on commands
such as `ls`.